pyproject.toml
[tool.poetry]
name = "imei-bot2"
version = "0.1.0"
description = ""
authors = ["Viktor"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
aiogram = "^3.17.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
